chaos

+++ 11_Linux/41_线程管理/pthread_join.c
@@ -6,7 +6,7 @@
 void* thread_function(void)
 {
 	int i;
-	printf("New thread, thread ID = %u\n", pthread_self());
+	printf("New thread, thread ID = %u\n", (unsigned)pthread_self());
 	for (i = 0; i < 15; i++)
 		printf("This is the new thread.\n");
 }
@@ -16,7 +16,7 @@ int main(void)
 	pthread_t thread_id;
 	int i, ret;
 	
-	printf("Main thread, thread ID = %u\n", pthread_self());
+	printf("Main thread, thread ID = %u\n", (unsigned)pthread_self());
 	ret = pthread_create(&thread_id, NULL, (void*)thread_function, NULL);
 	if (ret != 0) {
 		printf("Create thread error!\n");
